+++ .gitignore
@@ -7,3 +7,5 @@
 settings.json
 
 node_modules/
+
+npm-debug.log
