+++ imports/ui/components/elements/Dialog.js
@@ -2,7 +2,7 @@ import React, {Component, PropTypes} from 'react';
 import classNames from 'classnames';
 
 // components
-import Button from '../Button';
+import Button from './Button';
 
 export class Dialog extends Component {
