--- imports/ui/components/netsuite-builder/Dropdown.js
@@ -1,20 +1,0 @@
-import React, {Component} from 'react';
-
-export class Dropdown extends Component {
-  render() {
-    return (
-      <div class="btn-group">
-        <button type="button" class="btn btn-default dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-          Action <span class="caret"></span>
-        </button>
-        <ul class="dropdown-menu">
-          <li><a href="#">Action</a></li>
-          <li><a href="#">Another action</a></li>
-          <li><a href="#">Something else here</a></li>
-          <li role="separator" class="divider"></li>
-          <li><a href="#">Separated link</a></li>
-        </ul>
-      </div>
-    );
-  }
-}
