+++ imports/ui/pages/examples/ConditionBuilderTree.js
@@ -1,12 +1,11 @@
 import React, {Component, PropTypes} from 'react';
 
 // components
-import Button from '../components/Button';
-import QueryBuilder from '../components/conditions-builder';
+import Button from '../../components/elements/Button';
+import {QueryBuilder} from '../../components/conditions-builder';
 
 // fields
 import Fields from '/imports/api/fields/custom/fields';
-import Operators from '/imports/api/fields/operators';
 import Condition from '/imports/api/expression/condition';
