+++ imports/utils/index.js
@@ -1,6 +1,6 @@
 import {SimpleSchema} from 'meteor/aldeed:simple-schema';
 import {check} from 'meteor/check';
-import classification from '/imports/ui/pages/classification';
+import classification from '/imports/ui/pages/examples/classification';
 
 export const IDValidator = {
   _id: {
